refactor: create a separate file for bank account
- Initialise the exports in `models/__init__.py`
- Change import path for Bank in `routers/bank.py`

The reason to split these is because bank and bank account will have
their separate route, which means a separate file for routes and
schemas for both of them. So, it would make sense if the models were
also separated, making it more clear that these two are different models
and all that.